+++ Flight_Instruments_Display/TouchGFX/gui/include/gui/screen1_screen/Screen1View.hpp
@@ -14,8 +14,8 @@ public:
 
     virtual void Update_Speed_Text(uint8_t value);
 	virtual void Update_Altitude_Text(uint16_t value);
-	virtual void Update_Pitch_Text(uint8_t value);
-	virtual void Update_Roll_Text(uint8_t value);
+	virtual void Update_Pitch_Text(float value);
+	virtual void Update_Roll_Text(float value);
 
 protected:
 };
