+++ Flight_Instruments_Display/TouchGFX/gui/src/screen1_screen/Screen1Presenter.cpp
@@ -27,11 +27,11 @@ void Screen1Presenter::Update_Altitude_Text(uint16_t value){
 	view.Update_Altitude_Text(value);
 
 }
-void Screen1Presenter::Update_Pitch_Text(uint8_t value){
+void Screen1Presenter::Update_Pitch_Text(float value){
 	view.Update_Pitch_Text(value);
 
 }
-void Screen1Presenter::Update_Roll_Text(uint8_t value){
+void Screen1Presenter::Update_Roll_Text(float value){
 	view.Update_Roll_Text(value);
 
 }
